# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ishu - India's Premier Education, Tools & News Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  whatsappNumber?: string | null;
  role: UserRole;
  createdAt: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  whatsappNumber?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type ResultStatus = (typeof ResultStatus)[keyof typeof ResultStatus];

export const ResultStatus = {
  active: "active",
  upcoming: "upcoming",
  expired: "expired",
  coming_soon: "coming_soon",
} as const;

export interface Result {
  id: number;
  title: string;
  shortDescription: string;
  fullDescription?: string | null;
  category: string;
  state?: string | null;
  status: ResultStatus;
  totalPosts?: number | null;
  lastDate?: string | null;
  examDate?: string | null;
  requiredDocuments: string[];
  eligibility?: string | null;
  officialLink?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ResultListResponse {
  results: Result[];
  total: number;
  page: number;
  totalPages: number;
}

export type ResultStatsByCategoryItem = {
  category: string;
  count: number;
};

export type ResultStatsByStateItem = {
  state: string;
  count: number;
};

export interface ResultStats {
  totalActive: number;
  totalUpcoming: number;
  totalExpired: number;
  byCategory: ResultStatsByCategoryItem[];
  byState: ResultStatsByStateItem[];
}

export interface ResultCategory {
  id: number;
  name: string;
  slug: string;
  count: number;
}

export interface StateVacancy {
  state: string;
  count: number;
  activeCount: number;
}

export interface NewsArticle {
  id: number;
  title: string;
  shortDescription: string;
  content?: string | null;
  category: string;
  imageUrl?: string | null;
  language: string;
  author?: string | null;
  isTrending: boolean;
  viewCount: number;
  relatedNewsIds: number[];
  createdAt: string;
  updatedAt: string;
}

export interface NewsListResponse {
  articles: NewsArticle[];
  total: number;
  page: number;
  totalPages: number;
}

export interface NewsCategory {
  id: number;
  name: string;
  slug: string;
  count: number;
  color?: string | null;
}

export interface Tool {
  id: number;
  name: string;
  slug: string;
  description: string;
  category: string;
  icon?: string | null;
  isNew: boolean;
  usageCount: number;
}

export type ToolStatsPopularToolsItem = {
  name: string;
  usageCount: number;
};

export interface ToolStats {
  totalTools: number;
  totalUsage: number;
  popularTools: ToolStatsPopularToolsItem[];
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content?: string | null;
  category: string;
  tags: string[];
  imageUrl?: string | null;
  author: string;
  /** Reading time in minutes */
  readTime: number;
  isFeatured: boolean;
  viewCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface BlogListResponse {
  posts: BlogPost[];
  total: number;
  page: number;
  totalPages: number;
}

export interface BlogCategory {
  id: number;
  name: string;
  slug: string;
  count: number;
}

export interface ContactBody {
  name: string;
  email: string;
  phone?: string | null;
  subject: string;
  message: string;
}

export interface ContactResponse {
  message: string;
  id: number;
}

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  subject: string;
  message: string;
  createdAt: string;
}

export interface NotificationSubscribeBody {
  whatsappNumber: string;
  name: string;
  categories?: string[];
}

export interface NotificationResponse {
  message: string;
  subscriptionId: number;
}

export type AdminStatsRecentActivityItem = {
  type: string;
  description: string;
  createdAt: string;
};

export interface AdminStats {
  totalUsers: number;
  totalResults: number;
  totalNews: number;
  totalBlogs: number;
  totalContacts: number;
  totalNotificationSubscribers: number;
  recentActivity: AdminStatsRecentActivityItem[];
}

export interface AdminUsersResponse {
  users: User[];
  total: number;
  page: number;
  totalPages: number;
}

export type CreateResultBodyStatus =
  (typeof CreateResultBodyStatus)[keyof typeof CreateResultBodyStatus];

export const CreateResultBodyStatus = {
  active: "active",
  upcoming: "upcoming",
  expired: "expired",
  coming_soon: "coming_soon",
} as const;

export interface CreateResultBody {
  title: string;
  shortDescription: string;
  fullDescription?: string | null;
  category: string;
  state?: string | null;
  status: CreateResultBodyStatus;
  totalPosts?: number | null;
  lastDate?: string | null;
  examDate?: string | null;
  requiredDocuments: string[];
  eligibility?: string | null;
  officialLink?: string | null;
}

export interface CreateNewsBody {
  title: string;
  shortDescription: string;
  content?: string | null;
  category: string;
  imageUrl?: string | null;
  language?: string;
  author?: string | null;
  isTrending?: boolean;
}

export interface CreateBlogBody {
  title: string;
  slug: string;
  excerpt: string;
  content?: string | null;
  category: string;
  tags: string[];
  imageUrl?: string | null;
  author: string;
  readTime: number;
  isFeatured?: boolean;
}

export type ListResultsParams = {
  category?: string;
  state?: string;
  status?: ListResultsStatus;
  page?: number;
  limit?: number;
};

export type ListResultsStatus =
  (typeof ListResultsStatus)[keyof typeof ListResultsStatus];

export const ListResultsStatus = {
  active: "active",
  upcoming: "upcoming",
  expired: "expired",
  coming_soon: "coming_soon",
} as const;

export type ListNewsParams = {
  category?: string;
  language?: string;
  page?: number;
  limit?: number;
  search?: string;
};

export type ListToolsParams = {
  category?: string;
};

export type MergePdfBody = {
  files?: Blob[];
};

export type CompressPdfBodyQuality =
  (typeof CompressPdfBodyQuality)[keyof typeof CompressPdfBodyQuality];

export const CompressPdfBodyQuality = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type CompressPdfBody = {
  file?: Blob;
  quality?: CompressPdfBodyQuality;
};

export type SplitPdfBody = {
  file?: Blob;
  /** Page ranges like "1-3,5,7-9" */
  pages?: string;
};

export type ListBlogsParams = {
  category?: string;
  tag?: string;
  page?: number;
  limit?: number;
  search?: string;
};

export type ListAdminUsersParams = {
  page?: number;
  limit?: number;
};