# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ishu - India's Premier Education, Tools & News Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # RESULTS / VACANCIES
...
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news articles
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
            default: en
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: News list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsListResponse"
  /news/{id}:
    get:
      operationId: getNews
      tags: [news]
      summary: Get a single news article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found
...
  /blogs:
    get:
      operationId: listBlogs
      tags: [blogs]
      summary: List blog posts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Blog list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogListResponse"
  /blogs/{slug}:
    get:
      operationId: getBlog
      tags: [blogs]
      summary: Get a blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
...
  /notifications/subscribe:
    post:
      operationId: subscribeNotifications
      tags: [notifications]
      summary: Subscribe to WhatsApp/push notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSubscribeBody"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationResponse"

  # ADMIN
...
  /admin/news:
    post:
      operationId: createNews
      tags: [admin]
      summary: Create a news article (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
  /admin/blogs:
    post:
      operationId: createBlog
      tags: [admin]
      summary: Create a blog post (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        whatsappNumber:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        whatsappNumber:
          type: string
          nullable: true
      required: [name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user, message]
...
    NewsArticle:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        shortDescription:
          type: string
        content:
          type: string
          nullable: true
        category:
          type: string
        imageUrl:
          type: string
          nullable: true
        language:
          type: string
          default: en
        author:
          type: string
          nullable: true
        isTrending:
          type: boolean
          default: false
        viewCount:
          type: integer
          default: 0
        relatedNewsIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, shortDescription, category, language, isTrending, viewCount, relatedNewsIds, createdAt, updatedAt]
    NewsListResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/NewsArticle"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [articles, total, page, totalPages]
...
    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
          nullable: true
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: string
          nullable: true
        author:
          type: string
        readTime:
          type: integer
          description: Reading time in minutes
        isFeatured:
          type: boolean
          default: false
        viewCount:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, slug, excerpt, category, tags, author, readTime, isFeatured, viewCount, createdAt, updatedAt]
    BlogListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [posts, total, page, totalPages]
...
    NotificationSubscribeBody:
      type: object
      properties:
        whatsappNumber:
          type: string
        name:
          type: string
        categories:
          type: array
          items:
            type: string
      required: [whatsappNumber, name]
    NotificationResponse:
      type: object
      properties:
        message:
          type: string
        subscriptionId:
          type: integer
      required: [message, subscriptionId]
...
    CreateNewsBody:
      type: object
      properties:
        title:
          type: string
        shortDescription:
          type: string
        content:
          type: string
          nullable: true
        category:
          type: string
        imageUrl:
          type: string
          nullable: true
        language:
          type: string
          default: en
        author:
          type: string
          nullable: true
        isTrending:
          type: boolean
          default: false
      required: [title, shortDescription, category]
    CreateBlogBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
          nullable: true
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: string
          nullable: true
        author:
          type: string
        readTime:
          type: integer
        isFeatured:
          type: boolean
          default: false
      required: [title, slug, excerpt, category, tags, author, readTime]