# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ishu - India's Premier Education, Tools & News Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content?: string | null;
  category: string;
  tags: string[];
  imageUrl?: string | null;
  author: string;
  /** Reading time in minutes */
  readTime: number;
  isFeatured: boolean;
  viewCount: number;
  createdAt: Date;
  updatedAt: Date;
}