# lib/api-zod/src/generated/types/createNewsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ishu - India's Premier Education, Tools & News Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateNewsBody {
  title: string;
  shortDescription: string;
  content?: string | null;
  category: string;
  imageUrl?: string | null;
  language?: string;
  author?: string | null;
  isTrending?: boolean;
}