# lib/api-zod/src/generated/types/notificationSubscribeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ishu - India's Premier Education, Tools & News Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface NotificationSubscribeBody {
  whatsappNumber: string;
  name: string;
  categories?: string[];
}